+++ packages/zosfiles/src/methods/list/List.ts
@@ -76,7 +76,8 @@ export class List {
             };
         } catch (error) {
             this.log.error(error);
-            // Throw detailed error message with REST debugging info if that option is included
+
+            // Throw detailed error message with REST API debug information if messageResponse option is present
             if (options.messageResponse) {
                 throw error;
             } else {
@@ -199,7 +200,16 @@ export class List {
             };
         } catch (error) {
             this.log.error(error);
-            throw error;
+
+            // Throw detailed error message with REST API debug information if messageResponse option is present
+            if (options.messageResponse) {
+                throw error;
+            } else {
+                // Isolate the main error message to display
+                const shortMessage: string[] = error.message.split("message:");
+                // Throw short error message
+                throw new ImperativeError({msg: "Error listing resources", additionalDetails: shortMessage[shortMessage.length - 1].trim()});
+            }
         }
     }