+++ packages/zosfiles/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Zowe z/OS files SDK package will be documented in thi
 
 ## Recent Changes
 
-- Improve user friendliness of error for list members in a data set. [#935](https://github.com/zowe/zowe-cli/issues/935)
+- Improve user friendliness of errors for zosfiles package. [#935](https://github.com/zowe/zowe-cli/issues/935)
 
 ## `6.32.1`